move the add-ons saga to core
